requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
pydantic-settings==2.2.1
python-dotenv==1.0.1
openai==1.31.0
twilio==9.0.0
structlog==24.2.0